fix:constantディレクトリの階層変更 #91
