--- src/components/constants/outlines.ts
@@ -1,39 +1,0 @@
-export const outlines = [
-    {
-        category: "hack",
-        backgroundImagePath:
-            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
-        caption:
-            "Hackでは、主に「人の役に立つものをプログラミングで作ること」を行っています。\n製作物は、Webサイト・Bot・Windowsアプリ・スマホアプリ・OSなど様々なものを生み出しています。実用的に役に立つように見えなくても、人を笑わせることで役に立っている製作物もあります！\nまた、ハッカソンや競技プログラミング、CTFをしている部員もおり、プログラミングに関して活発な活動が行われています！",
-        redirectPath: "/web/artifact/hack",
-    },
-    {
-        category: "game",
-        backgroundImagePath:
-            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
-        caption: "",
-        redirectPath: "/web/artifact/game",
-    },
-    {
-        category: "2DCG",
-        backgroundImagePath:
-            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
-        caption:
-            "　2DCGでは、主にデジタルイラストや動画など、主に2Dで表現されるものを含みます。\n　デジタルイラストは、文字通りデジタルでイラストを描きます。アナログで絵を描くのとは違う表現が可能です。パソコンや専用機器がなくても、スマートフォンやタブレットで気軽に描くことができます！\n　動画制作は、複数種類の画像,音楽,動画に動きやエフェクトを付け、組み合わせることで 新たな映像作品やPVなどを作り上げることです。無料の動画編集ソフトを使って気軽に制作できます。",
-        redirectPath: "/web/artifact/cg2d",
-    },
-    {
-        category: "3DCG",
-        backgroundImagePath:
-            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
-        caption: "",
-        redirectPath: "/web/artifact/cg3d",
-    },
-    {
-        category: "MUSIC",
-        backgroundImagePath:
-            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
-        caption: "",
-        redirectPath: "/web/artifact/music",
-    },
-];
