+++ src/constants/outlines.ts
@@ -1,2 +1,41 @@
 export const aboutC3 =
     "　Composite Computer Club(略称C3)はコンピューターを用いて創作活動をしている部活です。活動内容は多岐にわたり、大きく分けるとGame、Hack、2DCG、3DCG、Musicの5つに分類されています。ですが、分野にとらわれず日々自由に活動をしている部員がたくさんいます！\n　初心者でも楽しく制作活動ができる環境を作っています！C3ではオープンキャンパスや工大祭などの学内イベントで作品を展示したり、ハッカソンなどの学外のイベントに参加している部員もいます。そして、部内展覧会を開催したり、「ToyBox」という部員専用投稿サイトを用いて、部内でも作品を見せ合うことのできる環境を作っています。\n　このようにC3ではコンピューターを用いて作品を作り、お互いに見せ合って刺激を得ることのできる部活になっています！\n　C3では多くの部員が様々な熱い思いを持って活動しています。もし、九工大情報工学部に入学することがあれば是非見学して、入部してもらえると嬉しいです！\n　当サイトや、先程挙げた「ToyBox」も部員が作成したもので、ここでは部員が作った創作物を体験することができます。C3に興味がある方、そうでない方も是非見ていってください！";
+export const outlines = [
+    {
+        category: "hack",
+        backgroundImagePath:
+            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
+        caption:
+            "Hackでは、主に「人の役に立つものをプログラミングで作ること」を行っています。\n製作物は、Webサイト・Bot・Windowsアプリ・スマホアプリ・OSなど様々なものを生み出しています。実用的に役に立つように見えなくても、人を笑わせることで役に立っている製作物もあります！\nまた、ハッカソンや競技プログラミング、CTFをしている部員もおり、プログラミングに関して活発な活動が行われています！",
+        redirectPath: "/web/artifact/hack",
+    },
+    {
+        category: "game",
+        backgroundImagePath:
+            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
+        caption: "",
+        redirectPath: "/web/artifact/game",
+    },
+    {
+        category: "2DCG",
+        backgroundImagePath:
+            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
+        caption:
+            "　2DCGでは、主にデジタルイラストや動画など、主に2Dで表現されるものを含みます。\n　デジタルイラストは、文字通りデジタルでイラストを描きます。アナログで絵を描くのとは違う表現が可能です。パソコンや専用機器がなくても、スマートフォンやタブレットで気軽に描くことができます！\n　動画制作は、複数種類の画像,音楽,動画に動きやエフェクトを付け、組み合わせることで 新たな映像作品やPVなどを作り上げることです。無料の動画編集ソフトを使って気軽に制作できます。",
+        redirectPath: "/web/artifact/cg2d",
+    },
+    {
+        category: "3DCG",
+        backgroundImagePath:
+            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
+        caption: "",
+        redirectPath: "/web/artifact/cg3d",
+    },
+    {
+        category: "MUSIC",
+        backgroundImagePath:
+            "https://img.freepik.com/free-photo/cute-kitten-staring-out-the-window-playful-curiosity-generative-ai_188544-12520.jpg",
+        caption: "",
+        redirectPath: "/web/artifact/music",
+    },
+];
